Complete Crud for DietPlan, testing on postman is yet to be done

+++ Crud/src/main/java/com/KnowIt/Gym_intellect_Crud/Controllers/Workout_joincontroller.java
@@ -42,7 +42,7 @@ public class Workout_joincontroller {
     }
 
     @GetMapping("/workouts-by-plan/{planId}")
-    public ResponseEntity<List<WorkoutJoinDTO>> getWorkoutsByPlan(@PathVariable int planId) {
+    public ResponseEntity<List<WorkoutJoinDTO>> getWorkoutsByPlan(@PathVariable long planId) {
         List<Workout_join> joins = workoutJoinService.getWorkoutsByPlan(planId);
         List<WorkoutJoinDTO> dtos = joins.stream()
             .map(join -> {
